requirements.txt
boto3==1.23.0
feedparser==6.0.8
Flask==2.1.2
Flask_Cors==3.0.10
python_dateutil==2.8.2
redis==4.3.1
requests==2.22.0
sentence_transformers==2.2.2
torch==1.10.2
